refactor: extract key board color scheme function
